refactor(demo): split cos_demo.cpp into module-based standalone demos
